refactor(privmx-endpoint): Move UserVerifierInterface interface from model to modules/core package
